l(z) is nearly there for LLS
